requirements.txt
streamlit
python-dotenv
openai-agents
requests
yfinance 